Linkie-BE/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
python-dotenv
pydantic
geoalchemy2
shapely
aioredis
